% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country-map.R
\name{counties}
\alias{counties}
\title{UK map at County/Unitary Authority level}
\usage{
counties()
}
\value{
A tibble with multipolygon geometry for each County/Unitary Authority.
}
\description{
Merges boundaries to County/Unitary Authority level from the unified dataset.
}
\examples{
counties()
}
