% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country-map.R
\name{country}
\alias{country}
\title{UK map at country level}
\usage{
country()
}
\value{
A tibble with multipolygon geometry for each country.
}
\description{
Dissolves administrative boundaries to country level.
}
\examples{
country()
}
