% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country-map.R
\name{lads}
\alias{lads}
\title{UK map at Local Authority Districts (LAD) (Local Government District (LGD) in Northern Ireland) level}
\usage{
lads()
}
\value{
A tibble with multipolygon geometry for each LAD/LGD.
}
\description{
Merges boundaries to LAD/LGD level from the Ward-level boundaries dataset.
}
\examples{
lads()
}
