% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country-map.R
\name{wpcs}
\alias{wpcs}
\title{UK map at Westminster Parliamentary Constituency (WPC) level}
\usage{
wpcs()
}
\value{
A tibble with multipolygon geometry for each LAD/LGD.
}
\description{
Merges boundaries to WPC level from the Ward-level boundaries dataset.
}
\examples{
wpcs()
}
